+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Graph's Dijkstra in case of no available path
 
 ## [1.26.0] - 2022-02-23